+++ packages/firebase_messaging/android/src/main/java/io/flutter/plugins/firebasemessaging/FirebaseMessagingPlugin.java
@@ -113,8 +113,12 @@ public class FirebaseMessagingPlugin extends BroadcastReceiver
                   channel.invokeMethod("onToken", task.getResult().getToken());
                 }
               });
+      result.success(null);
+    } else if ("getLaunchMessage".equals(call.method)) {
       if (registrar.activity() != null) {
-        sendMessageFromIntent("onLaunch", registrar.activity().getIntent());
+        Map<String, Object> message = this.getMessageFromIntent(registrar.activity().getIntent());
+        result.success(message);
+        return;
       }
       result.success(null);
     } else if ("subscribeToTopic".equals(call.method)) {
@@ -176,15 +180,14 @@ public class FirebaseMessagingPlugin extends BroadcastReceiver
     return res;
   }
 
-  /** @return true if intent contained a message to send. */
-  private boolean sendMessageFromIntent(String method, Intent intent) {
+  private Map<String, Object> getMessageFromIntent(Intent intent) {
     if (CLICK_ACTION_VALUE.equals(intent.getAction())
         || CLICK_ACTION_VALUE.equals(intent.getStringExtra("click_action"))) {
       Map<String, Object> message = new HashMap<>();
       Bundle extras = intent.getExtras();
 
       if (extras == null) {
-        return false;
+        return null;
       }
 
       Map<String, Object> notificationMap = new HashMap<>();
@@ -199,10 +202,20 @@ public class FirebaseMessagingPlugin extends BroadcastReceiver
 
       message.put("notification", notificationMap);
       message.put("data", dataMap);
-
-      channel.invokeMethod(method, message);
-      return true;
+      return message;
     }
-    return false;
+    return null;
+  }
+
+  /** @return true if intent contained a message to send. */
+  private boolean sendMessageFromIntent(String method, Intent intent) {
+    Map<String, Object> message = this.getMessageFromIntent(intent);
+
+    if (message == null) {
+      return false;
+    }
+
+    channel.invokeMethod(method, message);
+    return true;
   }
 }